+++ yash-syntax/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.11.0] - Unreleased
 
+### Added
+
+- The following functions are now `const`:
+    - `parser::lex::is_portable_name_char`
+    - `parser::lex::is_special_parameter_char`
+    - `parser::lex::is_single_char_name`
+- A new `syntax::Param` struct is introduced to represent a parameter in
+  parameter expansions (`syntax::TextUnit::RawParam` and `syntax::BracedParam`).
+  New enum types `syntax::SpecialParam` and `syntax::ParamType` are added to
+  represent the details of the parameter. Note that the former `syntax::Param`
+  struct is renamed to `syntax::BracedParam`.
+
 ### Changed
 
 - `syntax::Param` has been renamed to `syntax::BracedParam`.